mutation repository no longer store to query data source
